chore: gut metadb.py from 3,533 to 283 lines — drop DuckDB backend
Extract constants to lltk/tools/constants.py (DB_BLACKLIST,
CORPUS_SOURCE_RANKS, GENRE_AUTHORITY_CORPORA, normalize_title/author,
_jaro_winkler, _chunk_text_to_passages).

metadb.py is now a thin shim: re-exports from constants.py/vocabs.py,
provides prepare_corpus_df() as standalone function, and instantiates
the MetaDBCH singleton.

Skip DuckDB-based TestMetaDB tests. Update TestPrepareCorpusDf to use
standalone function instead of MetaDB class.
